Merge in MCU16CE/dspic33-wolfcrypt from ~C71887/dspic33-wolfcrypt:M16CRYPTO-244-dsPIC33A-LMS-Misra to develop

* commit '7bb88f4f76081cbec40c12373e2927c1cc37d0ee':
  M16CRYPTO-244 update comment for user_settings readme
  M16CRYPTO-244 update user settings name for LMS
  M16CRYPTO-244 revert uncomment of width and height for LMS demo
  M16CRYPTO-244 udpate misra violations missed
  M16CRYPTO-244 Misra check dsPIC33AK LMS demo

+++ dspic33ak512mps512/lms/dspic33ak512mps512-lms.X/crypto/dspic33a_cam_hash.c
@@ -37,7 +37,7 @@ int wc_InitSha256_ex(wc_Sha256* sha256, void* heap, int devId)
 
         status = Crypto_Hash_Sha_Init(&sha256->context, CRYPTO_HASH_SHA2_256, CRYPTO_HANDLER_HW_INTERNAL, 1);
 
-        if(status != CRYPTO_HASH_SUCCESS)
+        if(status != (int) CRYPTO_HASH_SUCCESS)
         {
             status = WC_HW_E;
         }
@@ -56,7 +56,7 @@ int wc_Sha256Update(wc_Sha256* sha256, const byte* data, word32 len)
 
         status = Crypto_Hash_Sha_Update(&sha256->context, localData, len);
 
-        if(status != CRYPTO_HASH_SUCCESS)
+        if(status != (int) CRYPTO_HASH_SUCCESS)
         {
             status = WC_HW_E;
         }
@@ -73,11 +73,11 @@ int wc_Sha256Final(wc_Sha256* sha256, byte* hash)
     {
         status = Crypto_Hash_Sha_Final(&sha256->context, hash);
 
-        if(status == CRYPTO_HASH_SUCCESS)
+        if(status == (int) CRYPTO_HASH_SUCCESS)
         {
             status = wc_InitSha256_ex(sha256, sha256->heap, INVALID_DEVID); /* reset state */
 
-            if(status != CRYPTO_HASH_SUCCESS)
+            if(status != (int) CRYPTO_HASH_SUCCESS)
             {
                 status = WC_HW_E;
             }
